Deleted old Navbar
