+++ src/index.css
@@ -10,6 +10,9 @@ body {
 	-webkit-font-smoothing: antialiased;
 	-moz-osx-font-smoothing: grayscale;
 	background-color: #121212;
+	max-width: 1600px;
+	margin-left: auto;
+	margin-right: auto;
 }
 
 code {
